yata-agent/pyproject.toml
[project]
name = "yata-agent"
version = "0.1.0"
description = "A simple math app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.13",
    "google-adk>=1.3.0",
    "google-api-python-client>=2.172.0",
    "google-auth-oauthlib[tool]>=1.2.2",
    "google-generativeai>=0.8.5",
    "markdown>=3.8",
    "openai>=1.88.0",
    "playwright>=1.52.0",
    "py-cord[voice]>=2.5.0",
    "pydub>=0.25.1",
    "pytest>=8.4.0",
    "python-dotenv==1.0.1",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
    "pytest-anyio",
    "pytest-asyncio>=0.23.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# ビルド対象のパッケージが src/calculator にあることを指定
packages = ["src/calculator"]

[tool.pytest.ini_options]
# テスト実行時に 'src' ディレクトリをPythonのパスに追加
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--continue-on-collection-errors"
asyncio_mode = "auto"

[project.scripts]
# ... existing code ...
